+++ p2panda-rs/src/test_utils/fixtures/fixtures.rs
@@ -114,72 +114,3 @@ pub fn v0_1_0_fixture() -> Fixture {
         entry: entry(message, seq_num(1), None, None)
     }
 }
-
-// More specific fixtures which can be used in the same way as above but also simply as helper functions
-// with no passed parameters returning a useful test value. This is especially needed in the templates where fixtures 
-// can't be injected in the usual way.
-
-pub mod defaults {
-    use rstest::fixture;
-
-    use crate::test_utils::fixtures;
-    use crate::hash::Hash;
-    use crate::message::Message;
-    use crate::entry::Entry;
-    
-    use super::{DEFAULT_HASH, DEFAULT_SCHEMA_HASH};
-
-    #[fixture]
-    pub fn default_some_hash() -> Option<Hash> {
-        fixtures::some_hash(DEFAULT_HASH)
-    }
-    
-    #[fixture]
-    pub fn default_message() -> Message {
-        fixtures::create_message(
-            fixtures::schema(DEFAULT_SCHEMA_HASH),
-            fixtures::fields(vec![("message", "Hello!")]),
-        )
-    }
-    
-    #[fixture]
-    pub fn default_update_message() -> Message {
-        fixtures::update_message(
-            fixtures::schema(DEFAULT_SCHEMA_HASH),
-            fixtures::hash(DEFAULT_HASH),
-            fixtures::fields(vec![("message", "Updated, hello!")]))
-    }
-    
-    #[fixture]
-    pub fn default_delete_message() -> Message {
-        fixtures::delete_message(
-            fixtures::schema(DEFAULT_SCHEMA_HASH),
-            fixtures::hash(DEFAULT_HASH),
-        )
-    }
-    
-    #[fixture]
-    pub fn first_entry() -> Entry {
-        fixtures::entry(default_message(), fixtures::seq_num(1), None, None)
-    }
-    
-    #[fixture]
-    pub fn entry_with_backlink() -> Entry {
-        fixtures::entry(
-            default_message(),
-            fixtures::seq_num(2),
-            default_some_hash(),
-            None,
-        )
-    }
-    
-    #[fixture]
-    pub fn entry_with_backlink_and_skiplink() -> Entry {
-        fixtures::entry(
-            default_message(),
-            fixtures::seq_num(13),
-            default_some_hash(),
-            default_some_hash(),
-        )
-    }   
-}
